10-k functions/loading-10k/requirements.txt
flask
google-cloud-storage
google-cloud-secret-manager
google-cloud-bigquery
requests
bs4